+++ scrapli/driver/async_driver.py
@@ -124,7 +124,7 @@ class AsyncScrape(ScrapeBase):
         self.logger.info(f"Connection to {self._initialization_args['host']} closed successfully")
 
     @property
-    def timeout_ops(self) -> int:
+    def timeout_ops(self) -> float:
         """
         Property for timeout_ops attribute
 
@@ -141,7 +141,7 @@ class AsyncScrape(ScrapeBase):
         return self._timeout_ops
 
     @timeout_ops.setter
-    def timeout_ops(self, timeout_value: int) -> None:
+    def timeout_ops(self, timeout_value: float) -> None:
         """
         Setter for timeout_ops attribute
 
@@ -149,7 +149,7 @@ class AsyncScrape(ScrapeBase):
         base driver class
 
         Args:
-            timeout_value: integer value to set as timeout_ops
+            timeout_value: float value to set as timeout_ops
 
         Returns:
             N/A  # noqa: DAR202
